Extract canvas frame source
