+++ src/sources/fabric.ts
@@ -1,7 +1,7 @@
 import * as fabric from 'fabric/node';
 import { type CanvasRenderingContext2D, createCanvas, ImageData } from 'canvas';
 import { boxBlurImage } from '../BoxBlur.js';
-import type { CreateFrameSourceOptions, FrameSource, CanvasLayer, CustomFabricFunctionCallbacks, Layer, OptionalPromise } from '../types.js';
+import type { CreateFrameSourceOptions, FrameSource, CustomFabricFunctionCallbacks, Layer, OptionalPromise } from '../types.js';
 
 export type FabricFrameSourceOptions<T> = CreateFrameSourceOptions<T> & { fabric: typeof fabric };
 export type FabricFrameSourceCallback<T> = (options: FabricFrameSourceOptions<T>) => OptionalPromise<CustomFabricFunctionCallbacks>;
@@ -79,26 +79,6 @@ export async function createFabricFrameSource<T extends Layer>(
   };
 }
 
-export async function createCustomCanvasFrameSource({ width, height, params }: Pick<CreateFrameSourceOptions<CanvasLayer>, "width" | "height" | "params">): Promise<FrameSource> {
-  const canvas = createCanvas(width, height);
-  const context = canvas.getContext('2d');
-
-  const { onClose, onRender } = await params.func(({ width, height, canvas }));
-
-  async function readNextFrame(progress: number) {
-    context.clearRect(0, 0, canvas.width, canvas.height);
-    await onRender(progress);
-    // require('fs').writeFileSync(`${new Date().getTime()}.png`, canvas.toBuffer('image/png'));
-    // I don't know any way to draw a node-canvas as a layer on a fabric.js canvas, other than converting to rgba first:
-    return canvasToRgba(context);
-  }
-
-  return {
-    readNextFrame,
-    // Node canvas needs no cleanup https://github.com/Automattic/node-canvas/issues/1216#issuecomment-412390668
-    close: onClose,
-  };
-}
 export type BlurImageOptions = {
   mutableImg: fabric.FabricImage,
   width: number,
